+++ firmware/controllers/rpm_calculator.c
@@ -79,7 +79,7 @@ static void shaftPositionCallback(ShaftEvents ckpEventType, int index) {
 			// / 4 because each cylinder sends a signal
 			// need to measure time from the previous non-skipped event
 
-			rpmState.rpm = 60000 * TICKS_IN_MS / engineConfiguration2->rpmMultiplier / diff;
+			rpmState.rpm = (int)(60000 * TICKS_IN_MS / engineConfiguration2->rpmMultiplier / diff);
 		}
 	}
 	rpmState.lastRpmEventTime = now;
